@@ -1,17 +1,18 @@
 apply plugin: 'com.android.application'
-
 apply plugin: "androidx.navigation.safeargs"
+apply plugin: 'com.google.gms.google-services'
 
 android {
-    compileSdkVersion 29
+    compileSdkVersion 30
     buildToolsVersion "29.0.2"
     defaultConfig {
         applicationId "br.com.hebaja.englishtrainingquizzes"
         minSdkVersion 19
-        targetSdkVersion 29
+        targetSdkVersion 30
         versionCode 1
         versionName "1.0"
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
+        multiDexEnabled true
     }
     buildTypes {
         release {
@@ -44,6 +45,16 @@ dependencies {
     implementation 'androidx.concurrent:concurrent-futures:1.1.0'
     implementation 'com.squareup.okhttp3:logging-interceptor:4.9.0'
 
+    implementation platform('com.google.firebase:firebase-bom:26.2.0')
+
+    implementation group: 'at.favre.lib', name: 'bcrypt', version: '0.9.0'
+    implementation 'com.google.android.gms:play-services-auth:19.0.0'
+
+    implementation 'com.facebook.android:facebook-android-sdk:[4,5)'
+
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'com.android.support:multidex:1.0.3'
+
     testImplementation 'junit:junit:4.13.1'
 
     debugImplementation 'androidx.fragment:fragment-testing:1.2.5'
@@ -57,6 +68,4 @@ dependencies {
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.3.0'
     androidTestImplementation 'androidx.test.espresso:espresso-contrib:3.3.0'
     androidTestImplementation 'androidx.test.espresso:espresso-intents:3.3.0'
-
-
 }
