@@ -1,5 +1,13 @@
 package br.com.hebaja.englishtrainingquizzes.retrofit;
 
+import br.com.hebaja.englishtrainingquizzes.retrofit.service.AveragesService;
+import br.com.hebaja.englishtrainingquizzes.retrofit.service.EmailExistsService;
+import br.com.hebaja.englishtrainingquizzes.retrofit.service.EmailUserPasswordResetService;
+import br.com.hebaja.englishtrainingquizzes.retrofit.service.FacebookIdTokenService;
+import br.com.hebaja.englishtrainingquizzes.retrofit.service.GoogleIdTokenService;
+import br.com.hebaja.englishtrainingquizzes.retrofit.service.EmailUserRegisterService;
+import br.com.hebaja.englishtrainingquizzes.retrofit.service.SaveExerciseService;
+import br.com.hebaja.englishtrainingquizzes.retrofit.service.UserEmailSignInService;
 import okhttp3.OkHttpClient;
 import okhttp3.logging.HttpLoggingInterceptor;
 import retrofit2.Retrofit;
@@ -7,9 +15,14 @@ import retrofit2.converter.gson.GsonConverterFactory;
 
 public class BaseRetrofit {
 
-    private final ExerciseService exerciseService;
-    private final UserService userService;
-
+    private final EmailUserRegisterService emailUserRegisterService;
+    private final EmailExistsService emailExistsService;
+    private final SaveExerciseService saveExerciseService;
+    private final AveragesService averagesService;
+    private final GoogleIdTokenService googleIdTokenService;
+    private final FacebookIdTokenService facebookIdTokenService;
+    private final UserEmailSignInService userEmailSignInService;
+    private final EmailUserPasswordResetService emailUserPasswordResetService;
 
     public BaseRetrofit() {
         HttpLoggingInterceptor interceptor = new HttpLoggingInterceptor();
@@ -19,17 +32,28 @@ public class BaseRetrofit {
                 .build();
 
         Retrofit retrofit = new Retrofit.Builder()
-                .baseUrl("http://192.168.1.100:8080/")
+//                .baseUrl("https://englishtrainingquizzes.herokuapp.com/")
+                .baseUrl("http://192.168.1.100:8080")
                 .client(client)
                 .addConverterFactory(GsonConverterFactory.create())
                 .build();
 
-        exerciseService = retrofit.create(ExerciseService.class);
-        userService = retrofit.create(UserService.class);
+        emailUserRegisterService = retrofit.create(EmailUserRegisterService.class);
+        emailExistsService = retrofit.create(EmailExistsService.class);
+        saveExerciseService = retrofit.create(SaveExerciseService.class);
+        averagesService = retrofit.create(AveragesService.class);
+        googleIdTokenService = retrofit.create(GoogleIdTokenService.class);
+        facebookIdTokenService = retrofit.create(FacebookIdTokenService.class);
+        userEmailSignInService = retrofit.create(UserEmailSignInService.class);
+        emailUserPasswordResetService = retrofit.create(EmailUserPasswordResetService.class);
     }
 
-    public ExerciseService getExerciseService() {
-        return exerciseService;
-    }
-    public UserService getUserService() { return userService; }
+    public EmailUserRegisterService getEmailUserRegisterService() { return emailUserRegisterService; }
+    public EmailExistsService getEmailExistsService() { return emailExistsService; }
+    public SaveExerciseService getSaveExerciseService() { return saveExerciseService; }
+    public AveragesService getAveragesService() { return averagesService; }
+    public GoogleIdTokenService getGoogleIdTokenService() { return googleIdTokenService; }
+    public FacebookIdTokenService getFacebookIdTokenService() { return facebookIdTokenService; }
+    public UserEmailSignInService getUserEmailSignInService() { return userEmailSignInService; }
+    public EmailUserPasswordResetService getEmailUserPasswordResetService() { return emailUserPasswordResetService; }
 }
