@@ -16,7 +16,6 @@ public class WarningViewModel extends AndroidViewModel {
 
     private MutableLiveData<WarningView> warningViewLiveData;
     private WarningView chosenOptionWarning;
-//    private int imageRes;
 
     public WarningViewModel(@NonNull Application application) {
         super(application);
@@ -24,23 +23,15 @@ public class WarningViewModel extends AndroidViewModel {
 
     public LiveData<WarningView> setVisibilityState(WarningView chosenOptionWarning) {
         checkIfDataIsNull();
-
         this.chosenOptionWarning = chosenOptionWarning;
         chosenOptionWarning.getCardView().setVisibility(View.INVISIBLE);
         warningViewLiveData.setValue(chosenOptionWarning);
         return warningViewLiveData;
     }
 
-    private void checkIfDataIsNull() {
-        if (warningViewLiveData == null) {
-            warningViewLiveData = new MutableLiveData<>();
-        }
-    }
-
     public LiveData<WarningView> setState(WarningView chosenOptionWarning, int imageRes, int tag, String answer, int color) {
         checkIfDataIsNull();
         this.chosenOptionWarning = chosenOptionWarning;
-//        this.imageRes = imageRes;
         chosenOptionWarning.setImageRes(imageRes);
         configureChosenOptionWarning(imageRes, tag, answer, color);
         chosenOptionWarning.getCardView().setVisibility(View.VISIBLE);
@@ -55,6 +46,7 @@ public class WarningViewModel extends AndroidViewModel {
     }
 
     public LiveData<WarningView> reset() {
+        checkIfDataIsNull();
         chosenOptionWarning.getCardView().setVisibility(View.INVISIBLE);
         warningViewLiveData.setValue(chosenOptionWarning);
 
@@ -71,6 +63,12 @@ public class WarningViewModel extends AndroidViewModel {
     private void setColorOfTextView(int color) {
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
             chosenOptionWarning.getTextView().setTextColor(getApplication().getColor(color));
+        }
+    }
+
+    private void checkIfDataIsNull() {
+        if (warningViewLiveData == null) {
+            warningViewLiveData = new MutableLiveData<>();
         }
     }
 }
